same valid for single/multi and moved func to helper
